fix: new db connect
