+++ packages/server/src/config/db.ts
@@ -1,29 +1,31 @@
-import { Client } from 'pg'
+import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
+import { Topic } from '../models/Topic'
+import { Comment } from '../models/Comment'
 
-const { POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT } =
-  process.env
+// const { POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT } =
+//   process.env
 
-export async function createClientAndConnect(): Promise<Client | null> {
+const { POSTGRES_DB, POSTGRES_PORT } = process.env
+
+const sequelizeOptions: SequelizeOptions = {
+  host: 'localhost',
+  port: Number(POSTGRES_PORT),
+  username: 'postgres',
+  password: 'postgres',
+  database: POSTGRES_DB,
+  dialect: 'postgres',
+}
+
+export const sequelize = new Sequelize(sequelizeOptions)
+
+sequelize.addModels([Topic, Comment])
+
+export async function dbConnect() {
   try {
-    const client = new Client({
-      user: POSTGRES_USER,
-      host: 'localhost',
-      database: POSTGRES_DB,
-      password: POSTGRES_PASSWORD,
-      port: Number(POSTGRES_PORT),
-    })
-
-    await client.connect()
-
-    const res = await client.query('SELECT NOW()')
-    console.log('  ➜ 🎸 Connected to the database at:', res?.rows?.[0].now)
-
-    // client.end()
-
-    return client
-  } catch (e) {
-    console.error(e)
+    await sequelize.authenticate() // Аутентификация в БД
+    await sequelize.sync() // Синхронизация базы данных
+    console.log('Connection has been established successfully')
+  } catch (error) {
+    console.error('Unable to connect to the database:', error)
   }
-
-  return null
 }
