--- packages/server/src/config/db_ext.ts
@@ -1,34 +1,0 @@
-import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
-import { Topic } from '../models/Topic'
-import { Comment } from '../models/Comment'
-
-const {
-  POSTGRES_USER,
-  POSTGRES_PASSWORD,
-  POSTGRES_DB,
-  POSTGRES_PORT,
-  POSTGRES_HOST,
-} = process.env
-
-const sequelizeOptions: SequelizeOptions = {
-  host: POSTGRES_HOST,
-  port: Number(POSTGRES_PORT),
-  username: POSTGRES_USER,
-  password: POSTGRES_PASSWORD,
-  database: POSTGRES_DB,
-  dialect: 'postgres',
-}
-
-export const sequelize = new Sequelize(sequelizeOptions)
-
-sequelize.addModels([Topic, Comment])
-
-export async function dbConnect() {
-  try {
-    await sequelize.authenticate()
-    await sequelize.sync()
-    console.log('Connection has been established successfully.')
-  } catch (error) {
-    console.error('Unable to connect to the database:', error)
-  }
-}
